pyproject.toml
[project]
name = "pesto-clipboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "dasbus>=1.7",
    "dbus-next>=0.2.3",
    "pillow>=12.0.0",
    "psutil>=7.1.3",
    "pydantic>=2.12.4",
    "pygobject>=3.54.5",
    "pystray>=0.19.5",
    "python-xlib>=0.33",
    "pywayland>=0.4.18",
    "sqlalchemy>=2.0.44",
    "sqlmodel>=0.0.27",
    "typer>=0.20.0",
    "watchdog>=6.0.0",
]

[tool.flake8]
per-file-ignores = [
    "main.py:E402"  # Allow gi.require_version before GTK imports
]
